Move DirectiveErrorCode and stack frame constants out of the model into fpy.bytecode.errors
