+++ src/fpy/codegen_llvm.py
@@ -17,7 +17,7 @@ except ImportError as exc:
         "with 'pip install fprime-fpy[wasm]'"
     ) from exc
 
-from fpy.model import DirectiveErrorCode
+from fpy.bytecode.errors import DirectiveErrorCode
 from fpy.state import CompileState
 from fpy.symbols import (
     BuiltinFuncSymbol,
